[BACKLOG-6547] - CDH55 secure shim: PUC: DB Table(s) Data Source cannot be created based on Secured Cloudera Impala JDBC connection
There are two issues here:
1. JDBC connection is not stored properly in BI Server repository
2. Incorrect error message displayed upon connection failure

The changes in this commit fix only the second issue.
It is caused by the fact that connect() method of Pentaho Hive1 driver threw exception when called by DriverManager with impala-url.
This happens since Hive1 JDBC driver in latest shims throws exception in constructor.
The fix is to return null instead if the caught exception has special flag indicating that this is not a real driver but a mocking stub.
The flag is propagated through SQLState property of SQLException class.