[BACKLOG-30638][BACKLOG-30731] Parquet In/Out support with S3/HCP (#1063)
Disabling FileSystem caching in the static FileSystem cache,
handling in the PvfsHadoopBridge.  This avoids reusing cached
filesystems that were created with different credentials in effect.

As part of this change, updated the bridge getFileStatus and listStatus
to return FileStatus with the pvfs paths set, rather than the path
of the underlying fs.  Formerly we were able to return the underlying
fs path, since those paths would be resolved against the cached fs.
Now we need to make sure actions on those paths still comes through
the pvfs bridge.

https://jira.pentaho.com/browse/BACKLOG-30638
https://jira.pentaho.com/browse/BACKLOG-30731