[PPP-4582] Use of Vulnerable Component: Components/oauth_client_library_for_java
PR Set for https://hv-eng.atlassian.net/browse/PPP-4582:
- https://github.com/pentaho/pdi-plugins-ee/pull/1290
- https://github.com/pentaho/pentaho-platform/pull/5676
-

Updated the gcs-connector versions to 3.0.2 where possible; it doesn't include gooogle-oauth-client at all.
- I had to add gcsio for apache driver so that it can be used in GcsConf for import com.google.cloud.hadoop.gcsio.GoogleCloudStorageFileSystem;

Dataproc1421 is stuck on Hadoop 2, so, we have to go with hadoop2-2.2.24 version, which has vulnerabilities pertaining to Hadoop 2.

All remaining instances of google oauth (including that from gcs-connector:jar:hadoop2-2.2.24 → google-oauth-client:jar:1.34.1) are now on 1.33.3+, whose indirect vulnerabilities pertain to guava, which is managed in its own dependency block.
- https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2023-2976
- https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2020-8908

I didn't see guava being pulled in on oauth anywhere, but I have excluded it in related, gateway-cloud-bindings dependencies to be sure that oauth doesn't pull it in and it is only ever intentionally added by us elsewhere.